==> src/0_define_legenda_sibcs .r <==

#Legend of soil classes according to SIBCS
#Taciara Zborowski Horst

#Summary
#This code creates a legend for soil classes in the second level category according
# to the Brazilian Soil Classification System (SIBCS).

#The legend includes the soil class names and their corresponding colors.
#The legend is saved as an RDS file for future use.



# Read the shapefile
soil_map <- sf::read_sf("data/ibge/pedo_area.shp")


# Prepare the color mapping dataframe
soil_legend2 <- data.frame(
  Class = c(
    "ARGISSOLO BRUNO-ACINZENTADO","ARGISSOLO ACINZENTADO","ARGISSOLO AMARELO","ARGISSOLO VERMELHO","ARGISSOLO VERMELHO-AMARELO",
    "CAMBISSOLO HÚMICO","CAMBISSOLO FLÚVICO","CAMBISSOLO HÁPLICO",
    "CHERNOSSOLO RÊNDZICO","CHERNOSSOLO EBÂNICO","CHERNOSSOLO ARGILÚVICO","CHERNOSSOLO HÁPLICO",
    "ESPODOSSOLO HUMILÚVICO","ESPODOSSOLO FERRILÚVICO","ESPODOSSOLO FERRI-HUMILÚVICO",

    "GLEISSOLO TIOMÓRFICO","GLEISSOLO SÁLICO","GLEISSOLO MELÂNICO","GLEISSOLO HÁPLICO",
    "LATOSSOLO BRUNO","LATOSSOLO AMARELO","LATOSSOLO VERMELHO","LATOSSOLO VERMELHO-AMARELO",
    "LUVISSOLO CRÔMICO","LUVISSOLO HÁPLICO",
    "NEOSSOLO LITÓLICO","NEOSSOLO FLÚVICO","NEOSSOLO REGOLÍTICO","NEOSSOLO QUARTZARÊNICO",
    "NITOSSOLO BRUNO","NITOSSOLO VERMELHO","NITOSSOLO HÁPLICO",
    "VERTISSOLO HÁPLICO","VERTISSOLO EBÂNICO","VERTISSOLO HIDROMÓRFICO",
    "PLINTOSSOLO HÁPLICO","PLINTOSSOLO ARGILÚVICO","PLINTOSSOLO PÉTRICO",
    "PLANOSSOLO HÁPLICO","PLANOSSOLO NÁTRICO",
    "ORGANOSSOLO HÁPLICO","ORGANOSSOLO FÓLICO","ORGANOSSOLO TIOMÓRFICO",
    "ÁREA URBANA","AFLORAMENTO","DUNAS","CORPO D'ÁGUA CONTINENTAL"
  ),
  Color = c(
# ARGISSOLO BRUNO-ACINZENTADO", "ARGISSOLO ACINZENTADO"
# ARGISSOLO AMARELO",  "ARGISSOLO VERMELHO", "ARGISSOLO VERMELHO-AMARELO",
    "#FFC8FA", "#FDF1F0", "#F1CCC8", "#F07F7F", "#FFA77F",
# "CAMBISSOLO HÚMICO","CAMBISSOLO FLÚVICO","CAMBISSOLO HÁPLICO",
    "#CFB680", "#EBDBBF","#D7C5A5",
# "CHERNOSSOLO RÊNDZICO","CHERNOSSOLO EBÂNICO","CHERNOSSOLO ARGILÚVICO","CHERNOSSOLO HÁPLICO",
    "#8E6856", "#AA8686", "#9C4A4E", "#A86A70",
# "ESPODOSSOLO HUMILÚVICO","ESPODOSSOLO FERRILÚVICO","ESPODOSSOLO FERRI-HUMILÚVICO",
    "#9AAABE", "#B6BEC5", "#9AA8BC",
# "GLEISSOLO TIOMÓRFICO","GLEISSOLO SÁLICO","GLEISSOLO MELÂNICO","GLEISSOLO HÁPLICO",
    "#6CA3CD", "#5EB4E6", "#78FCFA", "#B6D8EE",
# "LATOSSOLO BRUNO","LATOSSOLO AMARELO","LATOSSOLO VERMELHO","LATOSSOLO VERMELHO-AMARELO",
    "#A87000", "#FEC05C", "#F4B980", "#F7D1A6",
# "LUVISSOLO CRÔMICO","LUVISSOLO HÁPLICO",
    "#D49616", "#D7B27A",
# "NEOSSOLO LITÓLICO","NEOSSOLO FLÚVICO","NEOSSOLO REGOLÍTICO","NEOSSOLO QUARTZARÊNICO",
    "#969595", "#EEEBDC", "#CFCFCE", "#FFFE73",
# "NITOSSOLO BRUNO","NITOSSOLO VERMELHO","NITOSSOLO HÁPLICO", 
    "#68350A", "#A83800", "#734C00",
# "VERTISSOLO HÁPLICO","VERTISSOLO EBÂNICO","VERTISSOLO HIDROMÓRFICO",
    "#C0C091", "#868F72", "#9EAA85",
# "PLINTOSSOLO HÁPLICO","PLINTOSSOLO ARGILÚVICO","PLINTOSSOLO PÉTRICO",
    "#D6BAC9", "#E3B3CD", "#ECACCB", 
# "PLANOSSOLO HÁPLICO","PLANOSSOLO NÁTRICO",
    "#B5D6AE", "#89CAC7",
# "ORGANOSSOLO HÁPLICO","ORGANOSSOLO FÓLICO","ORGANOSSOLO TIOMÓRFICO",
    "#A7B3D5", "#8596C1", "#5E81A1",
# "ÁREA URBANA","AFLORAMENTO","DUNAS","CORPO D'ÁGUA CONTINENTAL"
    "red", "white", "white", "blue"
  ),
  stringsAsFactors = FALSE
)

# Print the corrected dataframe to verify
print(soil_legend2)
str(soil_legend2)

# Save
saveRDS(soil_legend2, file = "data/ibge/soil_classes_legend.rds")
